Fixed some things and added button)
In fact, I have fixed constraints for some labels. But I have added two
more unfixed labels with text…( however I have updated the code for the
action that will hide/unhide the things on the screen. and added one
uiview with ui image view on it connected to the button